requirements.txt
<<<<<<< HEAD
fastapi
uvicorn[standard]
tensorflow==2.12.0
pillow
numpy
python-multipart
sqlalchemy
pydantic
reportlab
requests
python-dotenv
geopy
gTTS
python-dateutil

=======
streamlit>=1.24
requests
Pillow
openai  # optional, only if you want GPT fallback
>>>>>>> 0c476bfea41943868b0a0183c9e10916f4988523
